Restructured the projects as it gets bigger, started building up the careers page.
