+++ client/src/ui/pages/job_application/small_blocks/Required.js
